Revise ADS1115.cpp to include ADS1015 functionality. Rename ADS1115.cpp to ADS1x15.cpp to reflect multi-chip compatability.
